spec: split internal.rs into internal/ directory
Split src/commands/spec/internal.rs (2063 lines) into internal/ with
7 files following the established pattern (assay: 10 files, scry: 7,
eval: 5):

- mod.rs: re-exports, DB_PATH constant
- types.rs: SpecType registry with 4 types + body templates (new)
- queries.rs: ReadySpec, BlockedSpec, SpecInfo, ListFilters (~602 lines)
- mutations.rs: promote/complete/abandon/pause/resume/block (~736 lines)
- split.rs: split_spec, split_spec_value (~212 lines)
- archive.rs: archive_spec, find_spec, resolve_spec_dir (~313 lines)
- queue.rs: next_spec, age, dep_counts, tag/tree helpers (~243 lines)

Pure restructuring — public API in mod.rs unchanged, all tests pass.
Tests moved to their respective files (3 archive tests, 1 queue test).

BumpType::from_spec_type() delegation deferred — release/mod.rs is in
the lib crate and cannot import from the binary crate's types.rs.
types.rs references BumpType correctly; delegation can happen when
types move to the lib crate.
